+++ src/net/post/HttpMessage.ts
@@ -1,8 +1,8 @@
 /**
  * ===========================
  * ParadigmCore: Blind Star
- * @name ExpressMessage.ts
- * @module src/net
+ * @name HttpMessage.ts
+ * @module src/net/post
  * ===========================
  *
  * @author Henry Harder
@@ -12,9 +12,13 @@
  * Simple class for creating and sending JSON messages using ExpressJS.
  */
 
+// Express.js webserver
 import * as exp from "express";
 
-export class Message {
+/**
+ * Represents response message for HTTP (post) API requests.
+ */
+export class HttpMessage {
 
     public static staticSendError(res: exp.Response, message: string, error: number): void {
         const json = {
